Use corrent datatype for array index loop

+++ src/math/Algebra.h
@@ -4,15 +4,15 @@ template <size_t rows, size_t cols>
 static void
 rowReduce(float (&matrix)[rows][cols]){
     //iterate down diagonol, zeroing up and down
-    for(int d=0; d<min(rows,cols); d++){
+    for(size_t d=0; d<min(rows,cols); d++){
         float m = matrix[d][d];
         if(m==0) continue;
         //iterate along rows that are not the row of interest
-        for(int r=0; r<rows; r++){
+        for(size_t r=0; r<rows; r++){
             if(r==d) continue;
             float factor = -matrix[r][d]/m;
             //iterate along row doing calculations
-            for(int i=d; i<cols; i++){
+            for(size_t i=d; i<cols; i++){
                 matrix[r][i] += factor*matrix[d][i];
             }
         }
